% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossed_join.R
\name{crossed_join}
\alias{crossed_join}
\title{Like \code{data.table::CJ()} but works on lists and
data frames like \code{expand.grid()}}
\usage{
crossed_join(X, sort = FALSE, unique = TRUE, log_limit = 8)
}
\arguments{
\item{X}{A list or data frame}

\item{sort}{Should the expansion be sorted? By default it is \code{FALSE}.}

\item{unique}{Should unique values across each column or list element
be taken? By default this is \code{TRUE}.}

\item{log_limit}{The maximum log10 limit for expanded number of rows.
Anything >= this results in an error.}
}
\value{
A data.table object
}
\description{
Like \code{data.table::CJ()} but works on lists and
data frames like \code{expand.grid()}
}
\examples{
library(timeplyr)
crossed_join(list(1:10, 11:20))
crossed_join(iris, sort = TRUE)
}
