% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{asc}
\alias{asc}
\alias{desc}
\title{Ascending & Descending order of vector}
\usage{
asc(x)

desc(x)
}
\arguments{
\item{x}{Vector.}
}
\value{
A numeric vector that orders in ascending or descending order. \cr
Useful in \code{dplyr::arrange()} or \code{farrange()}.
}
\description{
Ascending & Descending order of vector
}
\examples{
library(dplyr)
library(timeplyr)
library(nycflights13)
data(flights)

flights \%>\%
  fdistinct(dest) \%>\%
  farrange(desc(dest))
}
