% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edf.R
\name{edf}
\alias{edf}
\title{Grouped empirical cumulative distribution function applied to data}
\usage{
edf(x, g = NULL)
}
\arguments{
\item{x}{Numeric vector.}

\item{g}{Nuemric vector of group IDs.}
}
\description{
Like \code{dplyr::cume_dist(x)} and \code{ecdf(x)(x)}
but with added grouping functionality.
}
\examples{
library(timeplyr)
library(dplyr)
x <- sample(seq(-10, 10, 0.5), size = 10^2, replace = TRUE)
plot(sort(edf(x)))
all.equal(edf(x), ecdf(x)(x))
all.equal(edf(x), cume_dist(x))
\dontrun{
library(microbenchmark)
microbenchmark(edf(x), cume_dist(x))
}
g <- sample(letters[1:3], size = 10^2, replace = TRUE)

edf1 <- tibble(x, g) \%>\%
  mutate(edf = cume_dist(x),
         .by = g) \%>\%
  pull(edf)
edf2 <- edf(x, g = g)
all.equal(edf1, edf2)
}
