% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farrange.R
\name{farrange}
\alias{farrange}
\title{A \code{collapse} version of \code{dplyr::arrange()}}
\usage{
farrange(data, ..., .by = NULL, .by_group = FALSE)
}
\arguments{
\item{data}{A data frame.}

\item{...}{Variables to group by.}

\item{.by}{(Optional). A selection of columns to group by for this operation.
Columns are specified using tidy-select.}

\item{.by_group}{If \code{TRUE} the sorting will be first done by the group
variables.}
}
\description{
This is a fast and near-identical alternative to \code{dplyr::arrange()}
using the \code{collapse} package.

This function is faster when there are many groups or a large number of
rows.

You can use it as you would with dplyr,
e.g. \code{iris \%>\% farrange(Species, desc(Sepal.Length))}
}
