% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fduplicates.R
\name{fduplicates}
\alias{fduplicates}
\alias{fduplicates2}
\title{Find duplicate rows}
\usage{
fduplicates(
  data,
  ...,
  .keep_all = FALSE,
  .both_ways = FALSE,
  .add_count = FALSE,
  .keep_na = TRUE,
  .by = NULL,
  .cols = NULL
)

fduplicates2(
  data,
  ...,
  .keep_all = FALSE,
  .both_ways = FALSE,
  .add_count = FALSE,
  .keep_na = TRUE,
  .by = NULL
)
}
\arguments{
\item{data}{A data frame.}

\item{...}{Variables used to find duplicate rows.}

\item{.keep_all}{If \code{TRUE} then all columns of data frame are kept,
default is \code{FALSE}.}

\item{.both_ways}{If \code{TRUE} then duplicates and non-duplicate first instances
are retained. The default is \code{FALSE} which returns only duplicated rows.}

\item{.add_count}{If \code{TRUE} then a count column is added to denote the
number of duplicates (including first non-duplicate instance).
The naming convention of this column follows \code{dplyr::add_count()}.}

\item{.keep_na}{If \code{FALSE} then rows with all \code{NA} values are removed.
The default is \code{TRUE}.}

\item{.by}{(Optional). A selection of columns to group by for this operation.
Columns are specified using tidy-select.}

\item{.cols}{(Optional) alternative to \code{...} that accepts
a named character vector or numeric vector.
If speed is an expensive resource, it is recommended to use this.}
}
\description{
This function works like \code{dplyr::distinct()} in its handling of
arguments and data-masking but returns duplicate rows.
In certain situations in can be much faster than \code{data \%>\% group_by() \%>\% filter(n() > 1)}
when there are many groups.
\code{fduplicates2()} returns the same output but uses a different
method which utilises joins and is written almost entirely using dplyr.
}
\examples{
library(dplyr)
library(timeplyr)
library(nycflights13)

# Duplicates across all columns
flights \%>\%
  fduplicates()
# Duplicate flights with the same tail number and departure time
flights \%>\%
  fduplicates(tailnum, dep_time)
# Can use tidyverse select notation
flights \%>\%
  fduplicates(across(contains("dep_time")), .keep_all = FALSE)
# Similar to janitor::get_dupes()
flights \%>\%
  fduplicates(tailnum, dep_time, .keep_all = FALSE, .add_count = TRUE)
# For every day were there multiple flights that departed at the same time?
flights \%>\%
  group_by(year, month, day) \%>\%
  fduplicates(dep_time, arr_time, .both_ways = TRUE)
}
