% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgroup_by.R
\name{fgroup_by}
\alias{fgroup_by}
\title{\code{collapse} version of \code{dplyr::group_by()}}
\usage{
fgroup_by(
  data,
  ...,
  .add = FALSE,
  order = TRUE,
  .by = NULL,
  .cols = NULL,
  .drop = TRUE
)
}
\arguments{
\item{data}{data frame.}

\item{...}{Variables to group by.}

\item{.add}{Should groups be added to existing groups?
Default is \code{FALSE}.}

\item{order}{Should groups be ordered? If \code{FALSE}
groups will be ordered based on first-appearance.}

\item{.by}{(Optional). A selection of columns to group by for this operation.
Columns are specified using \code{tidyselect}.}

\item{.cols}{(Optional) alternative to \code{...} that accepts
a named character vector or numeric vector.
If speed is an expensive resource, it is recommended to use this.}

\item{.drop}{Should unused factor levels be dropped? Default is \code{TRUE}.}
}
\value{
A \code{grouped_df}.
}
\description{
This works the exact same as \code{dplyr::group_by()} and typically
performs around the same speed but uses slightly less memory.
}
