% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.R
\name{fn}
\alias{fn}
\alias{gn}
\title{Alternative to \code{dplyr::n()}}
\usage{
fn(..., g = NULL, use.g.names = FALSE)

gn(..., g = NULL)
}
\arguments{
\item{...}{Template with which to calculate group sizes.
This can be left unused as long as g is not NULL.
If a data frame or matrix is supplied,
the number of rows by group \code{g} is calculated.
If a vector is supplied, the group-wise lengths are calculated.}

\item{g}{Object to be used as groups, passed directly to \code{collapse::GRP()}.}

\item{use.g.names}{If \code{TRUE} group names are added to the result as names.
This only applies to \code{fn()}.}
}
\value{
An integer vector of row sizes.
}
\description{
Alternative to \code{dplyr::n()}
}
\details{
\code{fn()} and \code{gn()} efficiently calculate numbers of rows
per group. They utilise \code{collapse::GRPN()} but the behaviour
aligns a bit more closely with the \code{collapse} statistical functions. \cr
\code{fn()} always returns the unique number of rows by group. \cr
\code{gn()} is the same as \code{fn()} but the output is expanded to
match the size and order of your data.
}
\examples{
library(timeplyr)
library(ggplot2)
fn(mpg)
fn(mpg, g = mpg$manufacturer)
gn(mpg)
gn(mpg, g = mpg$manufacturer)

library(data.table)
dt <- data.table(g = sample.int(10^5))
dt[, .N, by = g] # Usual DT method
dt[, list(g, N = fn(g = g))] # Using fn()
}
