% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnmiss.R
\name{fnmiss}
\alias{fnmiss}
\title{Fast grouped numbers of missing values}
\usage{
fnmiss(x, g = NULL, ...)
}
\arguments{
\item{x}{Vector or data frame.}

\item{g}{Object to be used for grouping,
passed directly to \code{collapse::GRP()}.}

\item{...}{Additional arguments passed onto \code{collapse::fsum}.}
}
\description{
A wrapper around \code{fsum(is.na(x))} to calculate
numbers of missing values by group.
}
