% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/froll.R
\name{frollsum2}
\alias{frollsum2}
\alias{frollmean2}
\title{Rolling sum/mean - \bold{Deprecated}}
\usage{
frollsum2(
  x,
  n = length(x),
  partial = TRUE,
  weights = NULL,
  fill = NA,
  algo = c("fast", "exact"),
  na.rm = FALSE,
  hasNA = NA
)

frollmean2(
  x,
  n = length(x),
  partial = TRUE,
  weights = NULL,
  fill = NA,
  algo = c("fast", "exact"),
  na.rm = FALSE,
  hasNA = NA
)
}
\arguments{
\item{x}{Numeric vector.}

\item{n}{Rolling window size, default is \code{length(x)}.}

\item{partial}{Should calculations be done using partial windows?
Default is \code{TRUE}.}

\item{weights}{Importance weights. Must be the same length as x.
Currently, no normalisation of weights occurs.}

\item{fill}{Numeric value to pad by for values in incomplete window.}

\item{algo}{Character, default "fast". When set to "exact", then slower algorithm is used.
See \code{?data.table::frollmean} details for more information.}

\item{na.rm}{Should missing values be removed when calculating window? Defaults to \code{FALSE}.}

\item{hasNA}{If it is known that x contains \code{NA} then setting to \code{TRUE} will speed up.}
}
\description{
This function is deprecated, use \code{roll_mean} and \code{roll_sum} instead.
}
\examples{
library(timeplyr)

x <- 1:10
frollsum2(x) # Simple rolling total
frollmean2(x) # Simple moving average
frollsum2(x, n = 3)
frollmean2(x, n = 3)
frollsum2(x, n = 3, partial = FALSE)
frollmean2(x, n = 3, partial = FALSE)

# Plot of expected value of 'coin toss' over many flips
set.seed(42)
x <- sample(c(1, 0), 10^3, replace = TRUE)
ev <- frollmean2(x)
plot(1:length(x), ev)
abline(h = 0.5, lty = 2)
}
