% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frowid.R
\name{frowid}
\alias{frowid}
\title{These functions have been superseded by fastplyr functions}
\usage{
frowid(x, ascending = TRUE)
}
\arguments{
\item{x}{A vector, data frame or \code{GRP} object.}

\item{ascending}{When \code{ascending = TRUE} the row IDs are in
increasing order. When \code{ascending = FALSE} the row IDs are in
decreasing order.}
}
\value{
An integer vector of row IDs.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}

These functions can now be found in fastplyr. \cr
They are no longer recommended in this package and thus have been both
deprecated and superseded.
}
\details{
\code{frowid()} is like \code{data.table::rowid()} but uses
an alternative method for calculating row numbers.
When \code{x} is a collapse \code{GRP} object, it is considerably faster.
It is also faster for character vectors.
}
