% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fslice.R
\name{fslice}
\alias{fslice}
\title{Much faster version of \code{dplyr::slice()} when there are lots of groups.}
\usage{
fslice(data, ..., .by = NULL)
}
\arguments{
\item{data}{Data frame}

\item{...}{See \code{?dplyr::slice} for details.}

\item{.by}{(Optional). A selection of columns to group by for this operation.
Columns are specified using tidy-select.}
}
\description{
Much faster version of \code{dplyr::slice()} when there are lots of groups.
}
\examples{
library(timeplyr)
library(dplyr)
library(nycflights13)

\dontrun{
library(microbenchmark)
data(flights)
# This may take a couple of minutes
microbenchmark(fslice(flights, 1:3, .by = c(origin, dest, tailnum)),
               slice(flights, 1:3, .by = c(origin, dest, tailnum)),
               times = 3)
microbenchmark(fslice(flights, 1:3, .by = everything()),
               slice(flights, 1:3, .by = everything()),
               times = 3)
}
}
