% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcd.R
\name{gcd}
\alias{gcd}
\title{Greatest common divisor}
\usage{
gcd(x, tol = sqrt(.Machine$double.eps), na_rm = TRUE)
}
\arguments{
\item{x}{A \link{numeric} vector.}

\item{tol}{Tolerance. Be careful as this must
be a single positive number strictly less than 1.}

\item{na_rm}{If \code{TRUE} the default, \code{NA} values are ignored.}
}
\value{
A number representing the GCD.
}
\description{
Fast greatest common divisor using the Euclidean algorithm.
}
\examples{
library(timeplyr)
library(bench)
\dontshow{
.n_dt_threads <- data.table::getDTthreads()
.n_collapse_threads <- collapse::get_collapse()$nthreads
data.table::setDTthreads(threads = 2L)
collapse::set_collapse(nthreads = 1L)
}
gcd(c(5, 25))
mark(gcd(c(5, 25)))

x <- rnorm(10^6)
mark(gcd(x, tol = 1e-10))
\dontshow{
data.table::setDTthreads(threads = .n_dt_threads)
collapse::set_collapse(nthreads = .n_collapse_threads)
}
}
