% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmin.R
\name{gmin}
\alias{gmin}
\alias{gmax}
\title{A grouped version of \code{pmin()} and \code{pmax()}}
\usage{
gmin(x, g = NULL, na.rm = TRUE)

gmax(x, g = NULL, na.rm = TRUE)
}
\arguments{
\item{x}{An atomic vector.}

\item{g}{An integer vector of group IDs.}

\item{na.rm}{Should \code{NA} values be removed? Default is \code{TRUE}.}
}
\description{
This is like \code{pmin()}/\code{pmax()} in the sense that it returns
a vector of min/max values the same length as x.
It doesn't operate over multiple vectors like \code{pmin()} and \code{pmax()} do.
}
\examples{
library(timeplyr)
library(dplyr)
iris \%>\%
  add_group_id(Species, .name = "g") \%>\%
  mutate(min = gmin(Sepal.Length, g = g),
         max = gmax(Sepal.Length, g = g)) \%>\%
  # The below is equivalent to above
  mutate(min2 = min(Sepal.Length),
         max2 = max(Sepal.Length), .by = Species) \%>\%
  distinct(Species, min, max, min2, max2)
}
