% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_collapse.R
\name{group_collapse}
\alias{group_collapse}
\title{These functions have been superseded by fastplyr functions}
\usage{
group_collapse(
  data,
  ...,
  order = TRUE,
  sort = FALSE,
  ascending = TRUE,
  .by = NULL,
  .cols = NULL,
  id = TRUE,
  size = TRUE,
  loc = TRUE,
  start = TRUE,
  end = TRUE,
  .drop = df_group_by_drop_default(data)
)
}
\arguments{
\item{data}{A data frame or vector.}

\item{...}{Additional groups using tidy \code{data-masking} rules. \cr
To specify groups using \code{tidyselect}, simply use the \code{.by} argument.}

\item{order}{Should the groups be ordered?
\bold{THE PHYSICAL ORDER OF THE DATA IS NOT CHANGED.} \cr
When order is \code{TRUE} (the default) the group IDs will be ordered but not sorted.
If \code{FALSE} the order of the group IDs will be based on first appearance.}

\item{sort}{Should the data frame be sorted by the groups?}

\item{ascending}{Should groups be ordered in ascending order?
Default is \code{TRUE} and only applies when \code{order = TRUE}.}

\item{.by}{Alternative way of supplying groups using \code{tidyselect} notation.
This is kept to be consistent with other functions.}

\item{.cols}{(Optional) alternative to \code{...} that accepts
a named character vector or numeric vector.
If speed is an expensive resource, it is recommended to use this.}

\item{id}{Should group IDs be added? Default is \code{TRUE}.}

\item{size}{Should group sizes be added? Default is \code{TRUE}.}

\item{loc}{Should group locations be added? Default is \code{TRUE}.}

\item{start}{Should group start locations be added? Default is \code{TRUE}.}

\item{end}{Should group end locations be added? Default is \code{TRUE}.}

\item{.drop}{Should unused factor levels be dropped? Default is \code{TRUE}.}
}
\value{
A \code{tibble} of unique groups and an integer ID uniquely identifying each group.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}

These functions can now be found in fastplyr. \cr
They are no longer recommended in this package and thus have been both
deprecated and superseded.
}
