% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseq_len.R
\name{gseq_len}
\alias{gseq_len}
\title{Grouped version of \code{seq_len()}}
\usage{
gseq_len(length, g = NULL)
}
\arguments{
\item{length}{Sequence length.}

\item{g}{(Optional) vector of group IDs.}
}
\value{
An integer vector,
or double if \code{length > .Machine$integer.max}
}
\description{
This is an extension to \code{seq_len()}
that produces a running integer sequence for each group that
increments by 1 and starts at 1.
This is most useful for calculating row numbers.
It is similar to \code{data.table::rowid()}.
}
