% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsum.R
\name{gsum}
\alias{gsum}
\alias{gmean}
\alias{gmin}
\alias{gmax}
\title{Grouped statistical functions.}
\usage{
gsum(x, g = NULL, na.rm = TRUE, ...)

gmean(x, g = NULL, na.rm = TRUE, ...)

gmin(x, g = NULL, na.rm = TRUE, ...)

gmax(x, g = NULL, na.rm = TRUE, ...)
}
\arguments{
\item{x}{An atomic vector.}

\item{g}{An integer vector of group IDs.}

\item{na.rm}{Should \code{NA} values be removed? Default is \code{TRUE}.}

\item{...}{Additional parameters passed on to the collapse package
equivalents, \code{fsum()}, \code{fmean()}, \code{fmin()}, and \code{fmax()}.}
}
\description{
These functions are wrappers around the collapse equivalents
but always return a vector the same length as x. \cr
They all accept group IDs for grouped calculations.
}
\examples{
library(timeplyr)
library(dplyr)

# Dplyr
iris \%>\%
  mutate(mean = mean(Sepal.Length), .by = Species)

# Timeplyr
iris \%>\%
  mutate(mean = gmean(Sepal.Length, g = group_id(., Species)))

iris \%>\%
  add_group_id(Species, .name = "g") \%>\%
  mutate(min = gmin(Sepal.Length, g = g),
         max = gmax(Sepal.Length, g = g),
         sum = gsum(Sepal.Length, g = g),
         mean = gmean(Sepal.Length, g = g)) \%>\%
  # The below is equivalent to above
  mutate(min2 = min(Sepal.Length),
         max2 = max(Sepal.Length),
         sum2 = sum(Sepal.Length),
         mean2 = mean(Sepal.Length),
         .by = Species) \%>\%
  distinct(Species,
           min, min2,
           max, max2,
           sum, sum2,
           mean, mean2)
}
