% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gunique.R
\name{gunique}
\alias{gunique}
\alias{gduplicated}
\alias{gwhich_duplicated}
\alias{gsort}
\alias{gorder}
\title{Grouped \code{unique()}, \code{sort()} and \code{duplicated()}}
\usage{
gunique(x, g = NULL, sort = FALSE, order = TRUE, use.g.names = TRUE)

gduplicated(x, g = NULL, order = TRUE, all = FALSE)

gwhich_duplicated(x, g = NULL, order = TRUE, all = FALSE)

gsort(x, g = NULL, order = TRUE, use.g.names = TRUE)

gorder(x, g = NULL, order = TRUE)
}
\arguments{
\item{x}{A vector or data frame.}

\item{g}{Object used for grouping, passed directly to \code{collapse::GRP()}.\cr
This can for example be a vector or data frame.}

\item{sort}{Should the result be sorted? \cr
This only applies to \code{gunique()}.}

\item{order}{Should the groups be treated as ordered groups?
Default is \code{TRUE}.}

\item{use.g.names}{Should group names be used? Default is \code{TRUE}.}

\item{all}{If \code{TRUE}, \code{gduplicated()} returns all duplicated values,
including the first occurrence.}
}
\description{
These functions use \code{collapse} and are like the
\code{collapse} counterpart but differ in that they accept a group \code{g} argument
which allows for more flexible by-group sorting.
}
