% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_max.R
\name{min_max}
\alias{min_max}
\title{min-max summary}
\usage{
min_max(data, ..., .add = FALSE, sort = FALSE, .by = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{...}{Variables to calculate the min/max with.}

\item{.add}{Should the min/max variables be added to the dataset?
If \code{TRUE} the data is kept as is, and the min/max variables are simply
added as additional columns.
If \code{FALSE} (the default) a summary data frame of unique groups is returned,
containing the min/max values for each group.}

\item{sort}{Should the data frame be sorted by the groups?
The default is \code{FALSE}.}

\item{.by}{Alternative way of supplying groups using \code{tidyselect} notation.
This is kept to be consistent with other functions.}
}
\value{
A summary data frame containing the min/max values for each group.
}
\description{
This is equivalent to
\code{data \%>\% group_by(...) \%>\% summarise(across(..., list(min, max)))}
but much faster and more efficient.
}
\examples{
\dontrun{
library(timeplyr)
library(dplyr)
library(nycflights13)
flights <- flights \%>\%
  group_by(origin, dest, tailnum)
flights \%>\%
  min_max(dep_time, arr_time)
  bench::mark(e1 = flights \%>\%
                min_max(dep_time, arr_time, sort = TRUE),
              e2 = flights \%>\%
                summarise(across(c(dep_time, arr_time),
                                 list(~ min(.x, na.rm = TRUE),
                                      ~ max(.x, na.rm = TRUE)))),
              check = FALSE)
}
}
