% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_summary.R
\name{q_summary}
\alias{q_summary}
\title{Fast grouped quantiles.}
\usage{
q_summary(
  data,
  ...,
  probs = seq(0, 1, 0.25),
  type = 7L,
  pivot = c("wide", "long"),
  na.rm = TRUE,
  sort = FALSE,
  .by = NULL,
  .cols = NULL
)
}
\arguments{
\item{data}{A data frame.}

\item{...}{Variables used to calculate quantiles for.
Tidy data-masking applies.}

\item{probs}{Quantile probabilities.}

\item{type}{An integer from 5-9 specifying which algorithm to use.
See \code{\link{quantile}} for more details.}

\item{pivot}{Should data be pivoted wide or long? Default is \code{wide}.}

\item{na.rm}{Should \code{NA} values be removed? Default is \code{TRUE}.}

\item{sort}{Should groups be sorted? Default is \code{TRUE}.}

\item{.by}{(Optional). A selection of columns to group by for this operation.
Columns are specified using tidy-select.}

\item{.cols}{(Optional) alternative to \code{...} that accepts
a named character vector or numeric vector.
If speed is an expensive resource, it is recommended to use this.}
}
\value{
A \code{data.table} containing the quantile values for each group.
}
\description{
\code{collapse} and \code{data.table} are used for the calculations.
}
\examples{
library(timeplyr)
library(dplyr)
iris \%>\%
  q_summary(Sepal.Length, .by = Species)
}
