% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_factor.R
\name{quick_factor}
\alias{quick_factor}
\title{A faster version of \code{factor()}}
\usage{
quick_factor(
  x = integer(),
  levels = NULL,
  order = TRUE,
  na_exclude = TRUE,
  ordered = is.ordered(x)
)
}
\arguments{
\item{x}{A vector.}

\item{levels}{Optional factor levels.}

\item{order}{Should factor levels be sorted? Default is \code{TRUE}.
It typically is faster to set this to \code{FALSE}, in which case the levels
are sorted by order of first appearance.}

\item{na_exclude}{Should \code{NA} values be excluded from the factor levels?
Default is \code{TRUE}.}

\item{ordered}{Should the result be an ordered factor?}
}
\value{
A \code{factor}.
}
\description{
A faster version of \code{factor()}
}
\details{
This operates similarly to \code{collapse::qF()}.
The main difference internally is that \code{collapse::funique()} is used
and therefore s3 methods can be written for it. It also works for
\code{vctrs_rcrd} objects and therefore works for \code{time_interval} objects.

To drop unused factor levels, simply use \code{collapse::fdroplevels()}.
}
