% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{transform_year_month}
\alias{transform_year_month}
\alias{transform_year_quarter}
\alias{scale_x_year_month}
\alias{scale_x_year_quarter}
\alias{scale_y_year_month}
\alias{scale_y_year_quarter}
\title{Additional ggplot2 scales}
\usage{
transform_year_month()

transform_year_quarter()

scale_x_year_month(...)

scale_x_year_quarter(...)

scale_y_year_month(...)

scale_y_year_quarter(...)
}
\arguments{
\item{...}{Arguments passed to \code{scale_x_continuous} and \code{scale_y_continuous}.}
}
\value{
A ggplot2 scale or transform.
}
\description{
Additional scales and transforms for use with year_months and year_quarters
in ggplot2.
}
