% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_tbl.R
\name{seq_tbl}
\alias{seq_tbl}
\title{Vectorised sequence with ID}
\usage{
seq_tbl(from = 1L, to = 1L, by = 1L)
}
\arguments{
\item{from}{Vector of sequence starts.}

\item{to}{Vector of sequence ends.}

\item{by}{Vector of sequence increments.}
}
\value{
A \code{tibble} of 2 columns with the sequence IDs and sequences.
}
\description{
An efficient method of calculating multiple sequences
with unique IDs for each sequence.
}
