% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence2.R
\name{sequence2}
\alias{sequence2}
\title{Extension to \code{base::sequence()} that handles decimals}
\usage{
sequence2(nvec, from = 1, by = 1)
}
\arguments{
\item{nvec}{Vector of sequence lengths.}

\item{from}{Start of sequence(s).}

\item{by}{Unit increment of sequence(s).}
}
\description{
Extension to \code{base::sequence()} that handles decimals
}
\examples{
library(timeplyr)
sequence(1:3)
sequence2(1:3)

sequence(1:3, by = 0.1)
sequence2(1:3, by = 0.1)

sequence(c(3, 2), by = c(-0.1, 0.1))
sequence2(c(3, 2), by = c(-0.1, 0.1))
}
