% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_summarise.R
\name{stat_summarise}
\alias{stat_summarise}
\title{Fast grouped statistical summary for data frames.}
\usage{
stat_summarise(
  data,
  ...,
  stat = .stat_fns[1L],
  .names = NULL,
  na.rm = TRUE,
  .by = NULL,
  .cols = NULL,
  sort = TRUE,
  keep_class = TRUE
)
}
\arguments{
\item{data}{A data frame.}

\item{...}{Variables used to calculate the min/max values.
Tidy data-masking applies.}

\item{stat}{A character vector of statistical summaries to apply.
This can be one or more of the following: \cr
"n", "min", "max", "mean", "first", "last", "sd",
"var", "mode", "median", "nobs".}

\item{.names}{An optional glue specification passed to \code{stringr::glue()}.
If \code{.names = NULL}, then in the case of 1 variable, the function name
is used, i.e .names = \code{"{.fn}"} and in the case of \verb{>=1} variables,
\code{"{.col}_{.fn}"} is used.}

\item{na.rm}{Should \code{NA} values be removed? Default is \code{TRUE}.}

\item{.by}{(Optional). A selection of columns to group by for this operation.
Columns are specified using tidy-select.}

\item{.cols}{(Optional) alternative to \code{...} that accepts
a named character vector or numeric vector.
If speed is an expensive resource, it is recommended to use this.}

\item{sort}{Should groups be sorted? Default is \code{TRUE}.}

\item{keep_class}{Logical. If \code{TRUE} then the class of
the input data is retained.
If \code{FALSE}, which is sometimes faster, a \code{data.table} is returned.}
}
\value{
A summary data frame containing the summary values for each group.
}
\description{
\code{collapse} and \code{data.table} are used for the calculations.
}
\details{
\code{stat_summarise()} can apply multiple functions to multiple variables.

\code{stat_summarise()} is equivalent to \cr
\code{data \%>\% group_by(...) \%>\% summarise(across(..., list(...)))} \cr
but is faster and more efficient and accepts limited statistical functions.
}
\examples{
library(timeplyr)
library(dplyr)
iris \%>\%
  stat_summarise(Sepal.Length, Sepal.Width,
                 stat = c("n", "min", "max", "mean"),
                 .by = Species)
}
