% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_diff_gcd.R
\name{time_diff_gcd}
\alias{time_diff_gcd}
\title{Fast greatest common divisor of time differences}
\usage{
time_diff_gcd(
  x,
  time_type = c("auto", "duration", "period"),
  is_sorted = FALSE,
  tol = sqrt(.Machine$double.eps)
)
}
\arguments{
\item{x}{Time variable. \cr
Can be a \code{Date}, \code{POSIXt}, \code{numeric}, \code{integer}, \code{yearmon}, or \code{yearqtr}.}

\item{time_type}{If "auto", \code{periods} are used if \code{x} is a Date and
durations are used if \code{x} is a datetime.
Otherwise numeric differences are calculated.}

\item{is_sorted}{Is \code{x} sorted? If \code{FALSE}, the default,
\code{sort(unique(x))} is calculated, otherwise \code{unique(x)} is used.
The order of successive time differences may impact the result.}

\item{tol}{Tolerance of comparison. The time differences are rounded
using \code{digits = ceiling(abs(log10(tol)))} to try and avoid
precision issues.}
}
\value{
A double vector of length 1 or length 0 if \code{length(x)} is 0.
}
\description{
Fast greatest common divisor of time differences
}
\examples{
library(timeplyr)
time_diff_gcd(1:10)
time_diff_gcd(seq(0, 1, 0.2))

time_diff_gcd(time_seq(Sys.Date(), Sys.Date() + 100, time_by = "3 days"))
time_diff_gcd(time_seq(Sys.time(), Sys.time() + 10^3, time_by = "125 seconds"))

# Detects monthly granularity
dplyr::near(time_diff_gcd(time(AirPassengers)), 1/12)
}
