% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_units.R
\docType{data}
\name{.time_units}
\alias{.time_units}
\alias{.period_units}
\alias{.duration_units}
\alias{.extra_time_units}
\title{Time units}
\format{
An object of class \code{character} of length 21.

An object of class \code{character} of length 7.

An object of class \code{character} of length 11.

An object of class \code{character} of length 10.
}
\usage{
.time_units

.period_units

.duration_units

.extra_time_units
}
\description{
Time units
}
\keyword{datasets}
