% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_as_tibble.R
\name{ts_as_tibble}
\alias{ts_as_tibble}
\title{Turn \code{ts} into a \code{tibble}}
\usage{
ts_as_tibble(x, name = "time", value = "value", ...)
}
\arguments{
\item{x}{An object of class \code{ts}, \code{mts} or \code{xts}.}

\item{name}{Name of the output time column.}

\item{value}{Name of the output value column.}

\item{...}{additional variables passed onto tibble.}
}
\description{
While a method already exists in the \code{tibble} package,
this method works differently in 2 ways:
\itemize{
\item The time variable associated with the time-series is also returned.
\item The returned \code{tibble} is always in long format, even when the time-series
is multivariate.
}
}
\examples{
library(timeplyr)
library(ggplot2)
library(dplyr)

# Using the examples from ?ts

# Univariate
uts <- ts(cumsum(1 + round(rnorm(100), 2)),
          start = c(1954, 7), frequency = 12)
uts_tbl <- ts_as_tibble(uts)

## Multivariate
mts <- ts(matrix(rnorm(300), 100, 3), start = c(1961, 1), frequency = 12)
mts_tbl <- ts_as_tibble(mts)

uts_tbl \%>\%
  ggplot(aes(x = time, y = value)) +
  geom_line()

mts_tbl \%>\%
  ggplot(aes(x = time, y = value)) +
  geom_line() +
  facet_wrap("group", ncol = 1)
}
