% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_guess.R
\name{unit_guess}
\alias{unit_guess}
\title{Guess time unit and extract basic information.}
\usage{
unit_guess(x)
}
\arguments{
\item{x}{This can be 1 of 4 options:
\itemize{
\item A string, e.g. "7 days"
\item lubridate duration or period object, e.g. \code{days(1)} or \code{ddays(1)}.
\item A list, e.g. list("days" = 7)
\item A number, e.g. 5
}}
}
\value{
A list of length 3, including the unit, number and scale.
}
\description{
This is a simple R function to convert time units to a
common unit, with number and scale. \cr
See \code{.time_units} for a list of accepted
time units.
}
