% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which2.R
\name{which2}
\alias{which2}
\title{Efficient alternative to \code{which()}}
\usage{
which2(x, invert = FALSE)
}
\arguments{
\item{x}{A \link{logical} vector.}

\item{invert}{If \code{TRUE}, indices of values that are not \code{TRUE} are returned
(including \code{NA}). If \code{FALSE} (the default), only \code{TRUE} indices are returned.}
}
\value{
An unnamed integer vector.
}
\description{
Exactly the same as \code{which()} but is more memory efficient.
}
