% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ym.R
\name{ym}
\alias{ym}
\alias{yq}
\alias{YM}
\alias{YQ}
\title{Fast methods for creating year-months and year-quarters}
\usage{
ym(x)

yq(x)

YM(length = 0L)

YQ(length = 0L)
}
\arguments{
\item{x}{A \code{ym}, \code{yq}, or any other time-based object.}

\item{length}{Length of \code{ym} or \code{yq}.}
}
\description{
These are experimental methods for working with year-months and
year-quarters inspired by 'zoo' and 'tsibble'.
}
\details{
The biggest difference is that the underlying data is simply
the number of months/quarters since epoch. This makes integer
arithmetic very simple, and allows for fast coercion to \code{ym} and \code{yq}
from numeric vectors.

Printing method is also fast.
}
\examples{
library(timeplyr)
library(lubridate)

x <- ym(today())

# Adding 1 adds 1 month
x + 1
# Adding 12 adds 1 year
x + 12
# Sequence of yearmonths
x + 0:12

# If you unclass, do the same arithmetic, and coerce back to ym
# The result is always the same
ym(unclass(x) + 1)
ym(unclass(x) + 12)

# Initialise a ym or yq to the specified length
YM(0)
YQ(0)
YM(3)
YQ(3)

}
