% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossed_join.R
\name{crossed_join}
\alias{crossed_join}
\title{These functions have been superseded by fastplyr functions}
\usage{
crossed_join(
  X,
  sort = FALSE,
  unique = TRUE,
  as_dt = TRUE,
  strings_as_factors = FALSE
)
}
\arguments{
\item{X}{A list or data frame.}

\item{sort}{Should the expansion be sorted? By default it is \code{FALSE}.}

\item{unique}{Should unique values across each column or list element
be taken? By default this is \code{TRUE}.}

\item{as_dt}{Should result be a \code{data.table}?
By default this is \code{TRUE}. If \code{FALSE} a list is returned.}

\item{strings_as_factors}{Should strings be converted to factors before
expansion? The default is \code{FALSE} but setting to \code{TRUE} can offer
a significant speed improvement.}
}
\value{
A data.table or list object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}

These functions can now be found in fastplyr. \cr
They are no longer recommended in this package and thus have been both
deprecated and superseded.

Use \code{do.call(fastplyr::crossing, data)} to replicate \code{crossed_join(data)}.
}
