% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{asc}
\alias{asc}
\alias{desc}
\title{These functions have been superseded by fastplyr functions}
\usage{
asc(x)

desc(x)
}
\arguments{
\item{x}{Vector.}
}
\value{
A numeric vector that can be ordered in ascending or descending order. \cr
Useful in \code{dplyr::arrange()} or \code{fastplyr::f_arrange()}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}

These functions can now be found in fastplyr. \cr
They are no longer recommended in this package and thus have been both
deprecated and superseded.
}
