% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{asc}
\alias{asc}
\alias{desc}
\title{Ascending & Descending order of vector}
\usage{
asc(x)

desc(x)
}
\arguments{
\item{x}{Vector.}
}
\value{
A numeric vector that can be ordered in ascending or descending order. \cr
Useful in \code{dplyr::arrange()} or \code{farrange()}.
}
\description{
An alternate to \code{dplyr::desc()} which is much faster
for character vectors and factors.
}
\examples{
library(dplyr)
library(timeplyr)
library(nycflights13)
data(flights)

flights \%>\%
  fdistinct(dest) \%>\%
  farrange(desc(dest))
\dontrun{
library(bench)
mark(order(dplyr::desc(flights$tailnum)),
     order(timeplyr::desc(flights$tailnum)))
}
}
