% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farrange.R
\name{farrange}
\alias{farrange}
\title{These functions have been superseded by fastplyr functions}
\usage{
farrange(data, ..., .by = NULL, .by_group = FALSE, .cols = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{...}{Variables to arrange by.}

\item{.by}{(Optional). A selection of columns to group by for this operation.
Columns are specified using \code{tidyselect}.}

\item{.by_group}{If \code{TRUE} the sorting will be first done by the group
variables.}

\item{.cols}{(Optional) alternative to \code{...} that accepts
a named character vector or numeric vector.
If speed is an expensive resource, it is recommended to use this.}
}
\value{
A sorted \code{data.frame}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}

These functions can now be found in fastplyr. \cr
They are no longer recommended in this package and thus have been both
deprecated and superseded.
}
