% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farrange.R
\name{farrange}
\alias{farrange}
\title{A \code{collapse} version of \code{dplyr::arrange()}}
\usage{
farrange(data, ..., .by = NULL, .by_group = FALSE)
}
\arguments{
\item{data}{A data frame.}

\item{...}{Variables to group by.}

\item{.by}{(Optional). A selection of columns to group by for this operation.
Columns are specified using tidy-select.}

\item{.by_group}{If \code{TRUE} the sorting will be first done by the group
variables.}
}
\description{
This is a fast and near-identical alternative to \code{dplyr::arrange()}
using the \code{collapse} package.

\code{desc()} is like \code{dplyr::desc()} but works faster when
called directly on vectors. \cr
You can use \code{desc()} interchangeably with \code{dplyr} and \code{timeplyr}. \cr
\code{arrange(iris, desc(Species))} uses \code{dplyr}'s version. \cr
\code{farrange(iris, desc(Species))} uses \code{timeplyr}'s version.

\code{farrange()} is faster when there are many groups or a large number of
rows.
}
