% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdistinct.R
\name{fdistinct}
\alias{fdistinct}
\title{Find distinct rows}
\usage{
fdistinct(data, ..., .keep_all = FALSE, .by = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{...}{Variables used to find distinct rows.}

\item{.keep_all}{If \code{TRUE} then all columns of data frame are kept,
default is \code{FALSE}.}

\item{.by}{(Optional). A selection of columns to group by for this operation.
Columns are specified using tidy-select.}
}
\description{
Like \code{dplyr::distinct()} but faster when lots of
groups are involved.
}
