% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fexpand.R
\name{fexpand}
\alias{fexpand}
\alias{fcomplete}
\title{These functions have been superseded by fastplyr functions}
\usage{
fexpand(data, ..., expand_type = NULL, sort = FALSE, .by = NULL)

fcomplete(data, ..., expand_type = NULL, sort = FALSE, .by = NULL, fill = NA)
}
\arguments{
\item{data}{A data frame.}

\item{...}{Variables to expand.}

\item{expand_type}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Use \code{fastplyr::crossing()} and \code{fastplyr::nesting()}.}

\item{sort}{Logical. If \code{TRUE} expanded/completed variables are sorted.
The default is \code{FALSE}.}

\item{.by}{(Optional). A selection of columns to group by for this operation.
Columns are specified using tidy-select.}

\item{fill}{A named list containing value-name pairs
to fill the named implicit missing values.}
}
\value{
A \code{data.frame} of expanded groups.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}

These functions can now be found in fastplyr. \cr
They are no longer recommended in this package and thus have been both
deprecated and superseded.
}
