% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fselect.R
\name{fselect}
\alias{fselect}
\alias{frename}
\title{Fast \code{dplyr::select()}/\code{dplyr::rename()}}
\usage{
fselect(data, ..., .cols = NULL)

frename(data, ..., .cols = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{...}{Variables to select using \code{tidyselect}.
See \code{?dplyr::select} for more info.}

\item{.cols}{(Optional) faster alternative to \code{...} that accepts
a named character vector or numeric vector. \cr
No checks on duplicates column names are done when using \code{.cols}. \cr
If speed is an expensive resource, it is recommended to use this.}
}
\value{
A \code{data.frame} of selected columns.
}
\description{
\code{fselect()} operates the exact same way as \code{dplyr::select()} and
can be used naturally with \code{tidyselect} helpers.
It uses \code{collapse} to perform the actual selecting of variables and is
considerably faster than \code{dplyr} for selecting exact columns,
and even more so when supplying the \code{.cols} argument.
}
\examples{
\dontrun{
  library(timeplyr)
  library(dplyr)
  bench::mark(e1 = fselect(iris, Species, Sepal.Length),
              e2 = fselect(iris, .cols = c("Species", "Sepal.Length")),
              e3 = fselect(iris, all_of(c("Species", "Sepal.Length"))),
              e4 = select(iris, all_of(c("Species", "Sepal.Length"))))
}
}
