% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fselect.R
\name{fselect}
\alias{fselect}
\alias{frename}
\title{These functions have been superseded by fastplyr functions}
\usage{
fselect(data, ..., .cols = NULL)

frename(data, ..., .cols = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{...}{Variables to select using \code{tidy-select}.
See \code{?dplyr::select} for more info.}

\item{.cols}{(Optional) faster alternative to \code{...} that accepts
a named character vector or numeric vector. \cr
No checks on duplicates column names are done when using \code{.cols}. \cr
If speed is an expensive resource, it is recommended to use this.}
}
\value{
A \code{data.frame} of selected columns.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}

These functions can now be found in fastplyr. \cr
They are no longer recommended in this package and thus have been both
deprecated and superseded.
}
