% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fslice.R
\name{fslice}
\alias{fslice}
\alias{fslice_head}
\title{Faster \code{dplyr::slice()}/\code{dplyr::slice_head()}}
\usage{
fslice(data, ..., .by = NULL)

fslice_head(data, ..., n = 1, .by = NULL)
}
\arguments{
\item{data}{Data frame}

\item{...}{See \code{?dplyr::slice} for details.}

\item{.by}{(Optional). A selection of columns to group by for this operation.
Columns are specified using tidy-select.}

\item{n}{Number of rows.}
}
\description{
When there are lots of groups, \code{fslice()} and \code{fslice_head()} are much faster than
the dplyr equivalents.

Unlike \code{dplyr}, \code{fslice()} and \code{fslice_head()} always return a sliced data frame in the same
order it was given, similar to how \code{filter()} works.
}
\examples{
library(timeplyr)
library(dplyr)
library(nycflights13)

\dontrun{
library(microbenchmark)
data(flights)
# This may take a couple of minutes
microbenchmark(fslice(flights, 1:3, .by = c(origin, dest, tailnum)),
               slice(flights, 1:3, .by = c(origin, dest, tailnum)),
               times = 3)
microbenchmark(fslice(flights, 1:3, .by = everything()),
               slice(flights, 1:3, .by = everything()),
               times = 3)
}
}
