% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_id.R
\name{group_id}
\alias{group_id}
\alias{add_group_id}
\alias{row_id}
\alias{add_row_id}
\alias{group_order}
\alias{add_group_order}
\title{Fast group and row IDs}
\usage{
group_id(data, ..., order = TRUE, .by = NULL, as_qg = FALSE)

add_group_id(data, ..., order = TRUE, .by = NULL, .name = NULL, as_qg = FALSE)

row_id(data, ..., .by = NULL)

add_row_id(data, ..., .by = NULL, .name = NULL)

group_order(data, ..., .by = NULL, decreasing = FALSE)

add_group_order(data, ..., .by = NULL, decreasing = FALSE, .name = NULL)
}
\arguments{
\item{data}{A data frame or vector.}

\item{...}{Additional groups using tidy select notation.}

\item{order}{Should the groups be ordered?
\bold{THE PHYSICAL ORDER OF THE DATA IS NOT CHANGED.} \cr
When order is \code{TRUE} (the default) the group IDs will be ordered but not sorted.
If \code{FALSE} the order of the group IDs will be based on first appearance.}

\item{.by}{Alternative way of supplying groups using tidy
select notation. This is kept to be consistent with other functions.}

\item{as_qg}{Should the group IDs be returned as a
collapse "qG" class? The default (\code{FALSE}) always returns
an integer vector.}

\item{.name}{Name of the added group ID column which should be a
character vector of length 1.
If \code{NULL} then a column named "group_id" will be added,
and if one already exists, a unique name will be used.}

\item{decreasing}{Should the group order be increasing or decreasing?
The default is \code{FALSE} and is currently only implemented in \code{group_order()}.}
}
\description{
\itemize{
\item \code{group_id()} returns an integer vector of group IDs the same size as the data.
\item \code{add_group_id()} adds an integer column of group IDs.
\item \code{row_id()} returns an integer vector of row IDs.
\item \code{add_row_id()} adds an integer column of row IDs.
\item \code{group_order()} returns the order of the groups.
\item \code{add_group_order()} adds an integer column of the order of the groups.
}
}
\examples{
library(timeplyr)
library(dplyr)
library(ggplot2)
group_id(iris) # No groups
group_id(iris, Species) # Species groups
iris \%>\%
  group_by(Species) \%>\%
  group_id() # Same thing
group_id(iris, where(is.numeric)) # Groups across numeric values

iris \%>\%
  add_group_id(Species) \%>\%
  distinct(Species, group_id)

mm_mpg <- mpg \%>\%
  select(manufacturer, model) \%>\%
  arrange(desc(pick(everything())))

# Sorted/non-sorted groups
mm_mpg \%>\%
  add_group_id(everything(),
               .name = "sorted_id", order = TRUE) \%>\%
  add_group_id(manufacturer, model,
               .name = "not_sorted_id", order = FALSE) \%>\%
  distinct()
}
