% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_id.R
\name{group_id}
\alias{group_id}
\alias{add_group_id}
\alias{group_loc}
\title{Fast group IDs}
\usage{
group_id(
  data,
  ...,
  order = sort,
  .by = NULL,
  .name = NULL,
  as_qg = FALSE,
  sort = TRUE
)

add_group_id(
  data,
  ...,
  order = sort,
  .by = NULL,
  .name = NULL,
  as_qg = FALSE,
  sort = TRUE
)

group_loc(data, ..., order = TRUE, .by = NULL, sort = FALSE)
}
\arguments{
\item{data}{A data frame or vector.}

\item{...}{Additional groups using tidy select notation.}

\item{order}{Should the groups be ordered?
\bold{THE PHYSICAL ORDER OF THE DATA IS NOT CHANGED.} \cr
When order is \code{TRUE} (the default) the group IDs will be ordered but not sorted.
If \code{FALSE} the order of the group IDs will be based on first appearance.}

\item{.by}{Alternative way of supplying groups using tidy
select notation. This is kept to be consistent with other functions.}

\item{.name}{Name of the added group ID column which should be a
character vector of length 1.
If \code{NULL} then a column named "group_id" will be added,
and if one already exists, a unique name will be used.}

\item{as_qg}{Should the group IDs be returned as a
collapse "qG" class? The default (\code{FALSE}) always returns
an integer vector.}

\item{sort}{Should the data frame be sorted by the groups? \cr
This is only applicable to \code{group_loc()}. \cr
For \code{group_id()} this should \bold{NOT BE USED} and is currently being deprecated,
use \code{order} instead.}
}
\description{
\itemize{
\item \code{group_id()} returns an integer vector of group IDs.
\item \code{add_group_id()} adds an integer column of group IDs.
\item \code{group_loc()} returns a data frame of group locations in the data.
}
}
\examples{
library(timeplyr)
library(dplyr)
library(ggplot2)
group_id(iris) # No groups
group_id(iris, Species) # Species groups
iris \%>\%
  group_by(Species) \%>\%
  group_id() # Same thing
group_id(iris, where(is.numeric)) # Groups across numeric values

iris \%>\%
  add_group_id(Species) \%>\%
  distinct(Species, group_id)

mm_mpg <- mpg \%>\%
  select(manufacturer, model) \%>\%
  arrange(desc(pick(everything())))

# Sorted/non-sorted groups
mm_mpg \%>\%
  add_group_id(everything(),
               .name = "sorted_id", order = TRUE) \%>\%
  add_group_id(manufacturer, model,
               .name = "not_sorted_id", order = FALSE) \%>\%
  distinct()
}
