% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_rate.R
\name{growth_rate}
\alias{growth_rate}
\alias{rolling_growth_rate}
\title{Fast Growth Rates}
\usage{
growth_rate(x, na.rm = FALSE, log = FALSE, inf_fill = NULL)

rolling_growth_rate(
  x,
  n = length(x),
  partial = TRUE,
  na.rm = FALSE,
  log = FALSE,
  inf_fill = NULL
)
}
\arguments{
\item{x}{Numeric vector.}

\item{na.rm}{Should missing values be removed when calculating window? Defaults to \code{FALSE}.}

\item{log}{If \code{TRUE} then growth rates are calculated on the log-scale.}

\item{inf_fill}{Numeric value to replace \code{Inf} values with. Default behaviour is to keep \code{Inf} values.}

\item{n}{Rolling window size, default is \code{length(x)} which calculates a rolling growth rate of \code{x}
compared to the starting value.
When \code{n = 2}, this is equivalent to a rolling basic growth calculation.}

\item{partial}{Should rates be calculated outwith the window using partial windows? Default is \code{TRUE}.}
}
\description{
Calculate the rate of percentage change per unit (time) of a numeric vector.
This metric is equivalent to the compound annual growth rate (CAGR): doi.org/10.1002/smr.1847
It is assumed that \code{x} is a vector of values corresponding to regular units of time with no gaps.
The output is to be interpreted as the average percent change per unit time.
This can also be calculated using the geometric mean of percent changes.
}
