% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gunique.R
\name{gunique}
\alias{gunique}
\title{Grouped \code{unique()}}
\usage{
gunique(x, g = NULL, sort = FALSE)
}
\arguments{
\item{x}{A vector or data frame.}

\item{g}{Group IDs passed directly to \code{collapse::GRP()}.}

\item{sort}{Should unique values be sorted?
The default is \code{FALSE}.}
}
\description{
Return unique values by group.
When \code{g = NULL} this is equivalent to \code{unique()}.
For \code{lubridate} users, this is much faster than \code{unique()} for
intervals.

For a more tidyverse friendly version for data frames, see \code{?fdistinct}.
}
