% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gunique.R
\name{gunique}
\alias{gunique}
\title{Grouped \code{unique()}}
\usage{
gunique(x, g = NULL)
}
\arguments{
\item{x}{A vector or data frame.}

\item{g}{Group IDs passed directly to \code{collapse::GRP()}.
This can be a vector, list or data frame.}
}
\description{
Return unique values by group.
When \code{g = NULL} this is equivalent to \code{unique()}.
}
