% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interval_utils.R
\name{interval_start}
\alias{interval_start}
\alias{interval_width}
\alias{interval_end}
\alias{interval_count}
\title{Time interval utilities}
\usage{
interval_start(x)

interval_width(x)

interval_end(x)

interval_count(x)
}
\arguments{
\item{x}{A 'time_interval'.}

\item{na_rm}{Should \code{NA} values be removed? Default is \code{TRUE}.}

\item{...}{Additional arguments passed onto \code{time_diff}.}
}
\description{
Time interval utilities
}
\seealso{
\link{time_interval}
}
