% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_date.R
\name{is_date}
\alias{is_date}
\alias{is_datetime}
\alias{is_time}
\alias{is_time_or_num}
\title{Utility functions for checking if date or datetime}
\usage{
is_date(x)

is_datetime(x)

is_time(x)

is_time_or_num(x)
}
\arguments{
\item{x}{Time variable. \cr
Can be a \code{Date}, \code{POSIXt}, \code{numeric}, \code{integer}, \code{yearmon}, or \code{yearqtr}.}
}
\value{
A logical(1).
}
\description{
Utility functions for checking if date or datetime
}
