% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_dates.R
\name{missing_dates}
\alias{missing_dates}
\alias{n_missing_dates}
\title{Check for missing dates between first and last date}
\usage{
missing_dates(x)

n_missing_dates(x)
}
\arguments{
\item{x}{A date or datetime vector, or a data frame.}
}
\value{
A date vector if x is a vector, or a list if x is a \code{data.frame}.
}
\description{
Check for missing dates between first and last date
}
