% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_dates.R
\name{missing_dates}
\alias{missing_dates}
\alias{n_missing_dates}
\title{Check for missing dates}
\usage{
missing_dates(x, na.rm = TRUE)

n_missing_dates(x, na.rm = TRUE)
}
\arguments{
\item{x}{A date or datetime vector, list, or data frame.}

\item{na.rm}{Should \code{NA} values be removed? Default is \code{TRUE}.}
}
\description{
\code{missing_dates()} returns missing dates between
the first and last date of a date/datetime vector.
}
