% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset.R
\name{reset_timeplyr_options}
\alias{reset_timeplyr_options}
\title{Reset 'timeplyr' options}
\usage{
reset_timeplyr_options()
}
\value{
Resets the timeplyr global options (prefixed with \code{"timeplyr."}): \cr
roll_month & roll_dst.
}
\description{
Reset 'timeplyr' options
}
