% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset.R
\name{reset_timeplyr_options}
\alias{reset_timeplyr_options}
\title{Reset 'timeplyr' options}
\usage{
reset_timeplyr_options()
}
\value{
Resets the timeplyr global options (prefixed with \code{"timeplyr."}): \cr
time_type, roll_month, roll_dst, interval_style, and
interval_sub_formatter.
}
\description{
Reset 'timeplyr' options
}
\examples{
library(timeplyr)
options(timeplyr.interval_style = "start")
getOption("timeplyr.interval_style")
reset_timeplyr_options()
getOption("timeplyr.interval_style")

}
