% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_resolution.R
\name{resolution}
\alias{resolution}
\alias{granularity}
\title{Time resolution & granularity}
\usage{
resolution(x, ...)

granularity(x, ...)
}
\value{
A \link{timespan} object.
}
\description{
The definitions of resolution and granularity may evolve over time but
currently the resolution defines the smallest timespan
that differentiates two non-fractional instances in time.
The granularity defines the smallest common timespan.
}
