% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll_lag.R
\name{roll_lag}
\alias{roll_lag}
\alias{lag_seq}
\title{Fast rolling lag}
\usage{
roll_lag(x, lag = 1L, check = TRUE)

lag_seq(x, lag = 1L)
}
\arguments{
\item{x}{A vector.}

\item{lag}{A vector of lags to be applied to \code{x} on a rolling basis.}

\item{check}{Should safety checks be done for the specified lag  against \code{x}?

The default is \code{TRUE} but
you can set this to \code{FALSE} if you need this to be speed-performant.}
}
\description{
An efficient rolling lag.

\code{lag_seq()} is a helper that creates a safe vector of lags.
}
\details{
This function is simple and fast. It should generally be equivalent
to
\code{diag(matrix(unlist(data.table::shift(x, lag)), ncol = length(x)))}.
}
