% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll_na_fill.R
\name{roll_na_fill}
\alias{roll_na_fill}
\title{Fast grouped "locf" \code{NA} fill}
\usage{
roll_na_fill(x, g = NULL, fill_limit = NULL)
}
\arguments{
\item{x}{A vector}

\item{g}{An object use for grouping x
This may be a vector or data frame for example.}

\item{fill_limit}{(Optional) maximum number of consecutive NAs to fill
per \code{NA} cluster.}
}
\value{
A filled vector of \code{x} the same length as \code{x}.
}
\description{
A fast by-group method for
"last-observation-carried-forward" \code{NA} filling.
}
\examples{
library(timeplyr)
library(dplyr)
library(data.table)

words <- do.call(paste0,
                 do.call(expand.grid, rep(list(letters), 3)))
groups <- sample(words, size = 10^5, replace = TRUE)
x <- sample.int(10^2, 10^5, TRUE)
x[sample.int(10^5, 10^4)] <- NA

dt <- data.table(x, groups)

roll_na_fill(x, groups)
\dontrun{
library(zoo)
library(imputeTS)
library(runner)

  # Summary
# Latest version of vctrs with their vec_fill_missing
# Is by far the best for repeated filling and memory efficiency
# For low repetitions and large vectors, data.table is best

# For large numbers of repetitions (groups) and data
# that is sorted by groups
# timeplyr is fastest, but not memory efficient

# No groups
bench::mark(e1 = dt[, filled := timeplyr::roll_na_fill(x)]$filled,
            e2 = dt[, filled := data.table::nafill(x, type = "locf")]$filled,
            e3 = dt[, filled := vctrs::vec_fill_missing(x)]$filled,
            e4 = dt[, filled := zoo:::na.locf0(x)]$filled,
            e5 = dt[, filled := imputeTS::na_locf(x, na_remaining = "keep")]$filled,
            e6 = dt[, filled := runner::fill_run(x, run_for_first = FALSE)]$filled)
# With group
bench::mark(e1 = dt[, filled := timeplyr::roll_na_fill(x, groups)]$filled,
            e2 = dt[, filled := data.table::nafill(x, type = "locf"), by = groups]$filled,
            e3 = dt[, filled := vctrs::vec_fill_missing(x), by = groups]$filled)
# Data sorted by groups
setkey(dt, groups)
bench::mark(e1 = dt[, filled := timeplyr::roll_na_fill(x, groups)]$filled,
            e2 = dt[, filled := data.table::nafill(x, type = "locf"), by = groups]$filled,
            e3 = dt[, filled := vctrs::vec_fill_missing(x), by = groups]$filled)
}
}
