% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll_sum.R
\name{roll_sum}
\alias{roll_sum}
\alias{roll_mean}
\title{Fast by-group rolling sum/mean}
\usage{
roll_sum(
  x,
  window = length(x),
  g = NULL,
  partial = TRUE,
  weights = NULL,
  na.rm = TRUE,
  ...
)

roll_mean(
  x,
  window = length(x),
  g = NULL,
  partial = TRUE,
  weights = NULL,
  na.rm = TRUE,
  ...
)
}
\arguments{
\item{x}{Numeric vector.}

\item{window}{Rolling window size, default is \code{length(x)}.}

\item{g}{Grouping object passed directly to \code{collapse::GRP()}.
This can for example be a vector or data frame.}

\item{partial}{Should calculations be done using partial windows?
Default is \code{TRUE}.}

\item{weights}{Importance weights. Must be the same length as x.
Currently, no normalisation of weights occurs.}

\item{na.rm}{Should missing values be removed for the calculation?
The default is \code{TRUE}.}

\item{...}{Additional arguments passed to \code{data.table::frollmean} and
\code{data.table::frollsum}.}
}
\description{
An efficient method for rolling sum/mean for many groups.
}
\examples{
\dontrun{
library(timeplyr)
library(data.table)
library(bench)
df <- data.table(g = sample.int(10^5, 10^6, TRUE),
                 x = rnorm(10^6))
mark(e1 = df[, mean := frollmean(x, n = 7, align = "right", na.rm = FALSE), by = "g"]$mean,
     e2 = df[, mean := roll_mean(x, n = 7, g = get("g"), partial = FALSE, na.rm = FALSE)]$mean)
}
}
