% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence2.R
\name{sequence2}
\alias{sequence2}
\alias{seq_id}
\alias{seq_v}
\title{Extension to \code{base::sequence()} that handles decimals}
\usage{
sequence2(nvec, from = 1L, by = 1L)

seq_id(nvec)

seq_v(from = 1L, to = 1L, by = 1L)
}
\arguments{
\item{nvec}{Vector of sequence lengths.}

\item{from}{Start of sequence(s).}

\item{by}{Unit increment of sequence(s).}

\item{to}{End of sequence(s).}
}
\description{
\code{seq_v()} is an alternate specification that instead accepts
the arguments \code{from}, \code{to} and \code{by}.
\code{seq_id()} is a helper function to efficiently group
each sequence by returning unique IDs along the sequences. \cr
}
\examples{
library(timeplyr)
sequence(1:3)
sequence2(1:3)

sequence(1:3, by = 0.1)
sequence2(1:3, by = 0.1)

sequence(c(3, 2), by = c(-0.1, 0.1))
sequence2(c(3, 2), by = c(-0.1, 0.1))
}
