% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_summarise.R
\name{stat_summarise}
\alias{stat_summarise}
\title{Fast statistical summary for data frames.}
\usage{
stat_summarise(
  data,
  ...,
  stat = "mean",
  na.rm = TRUE,
  .by = NULL,
  .cols = NULL,
  sort = TRUE
)
}
\arguments{
\item{data}{A data frame.}

\item{...}{Variables used to calculate the min/max values.
Tidy data-masking applies.}

\item{stat}{A character vector of statistical summaries to apply.
This can be one or more of the following: \cr
"min", "max", "mean", "first", "last", "sd", "var", "mode", "median", "nobs".}

\item{na.rm}{Should \code{NA} values be removed? Default is \code{TRUE}.}

\item{.by}{(Optional). A selection of columns to group by for this operation.
Columns are specified using tidy-select.}

\item{.cols}{(Optional) alternative to \code{...} that accepts
a named character vector or numeric vector.
If speed is an expensive resource, it is recommended to use this.}

\item{sort}{Should groups be sorted? Default is \code{TRUE}.}
}
\value{
A summary data frame containing the summary values for each group.
}
\description{
\code{collapse} and \code{data.table} are used for the calculations.
}
\details{
\code{stat_summarise()} can apply multiple functions to multiple variables.

\code{stat_summarise()} is equivalent to \cr
\code{data \%>\% group_by(...) \%>\% summarise(across(..., list(...)))} \cr
but is faster and more efficient and accepts limited statistical functions.
}
\examples{
library(timeplyr)
library(dplyr)
iris \%>\%
  stat_summarise(Sepal.Length, Sepal.Width,
                 stat = c("min", "max", "mean"),
                 .by = Species)
}
