% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_summarise.R
\docType{data}
\name{stat_summarise}
\alias{stat_summarise}
\alias{.stat_fns}
\title{These functions have been superseded by fastplyr functions}
\format{
An object of class \code{character} of length 14.
}
\usage{
stat_summarise(
  data,
  ...,
  stat = .stat_fns[1:3],
  q_probs = NULL,
  na.rm = TRUE,
  sort = df_group_by_order_default(data),
  .count_name = NULL,
  .names = NULL,
  .by = NULL,
  .cols = NULL,
  inform_stats = TRUE,
  as_tbl = FALSE
)

.stat_fns
}
\arguments{
\item{data}{A data frame.}

\item{...}{Variables to apply the statistical functions to.
Tidy data-masking applies.}

\item{stat}{A character vector of statistical summaries to apply.
This can be one or more of the following: \cr
"n", "nmiss", "ndistinct", "min", "max", "mean", "first", "last", "sd",
"var", "mode", "median", "sum", "prop_complete".}

\item{q_probs}{(Optional) Quantile probabilities.
If supplied, \code{q_summarise()} is called and added to the result.}

\item{na.rm}{Should \code{NA} values be removed? Default is \code{TRUE}.}

\item{sort}{Should groups be sorted? Default is \code{TRUE}.}

\item{.count_name}{Name of count column, default is "n".}

\item{.names}{An optional glue specification passed to \code{stringr::glue()}.
If \code{.names = NULL}, then when there is 1 variable, the function name
is used, i.e \code{.names = "{.fn}"}, when there are multiple variables and
1 function, the variable names are used, i.e, \code{.names = "{.col}"}
and in the case of multiple variables and functions.
\code{"{.col}_{.fn}"} is used.}

\item{.by}{(Optional). A selection of columns to group by for this operation.
Columns are specified using tidy-select.}

\item{.cols}{(Optional) alternative to \code{...} that accepts
a named character vector or numeric vector.
If speed is an expensive resource, it is recommended to use this.}

\item{inform_stats}{Should available stat functions be displayed
at the start of each session? Default is \code{TRUE}.}

\item{as_tbl}{Should the result be a \code{tibble}? Default is \code{FALSE}.}
}
\value{
A summary \code{data.table} containing the summary values for each group.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}

These functions can now be found in fastplyr. \cr
They are no longer recommended in this package and thus have been both
deprecated and superseded.
}
\keyword{datasets}
