% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_arithmetic.R
\name{time_add}
\alias{time_add}
\title{Add/subtract timespans to dates and date-times}
\usage{
time_add(
  x,
  timespan,
  roll_month = getOption("timeplyr.roll_month", "preday"),
  roll_dst = getOption("timeplyr.roll_dst", c("NA", "pre"))
)
}
\arguments{
\item{x}{Time vector. \cr
E.g. a \code{Date}, \code{POSIXt}, \code{numeric} or any time-based vector.}

\item{timespan}{\link{timespan}.}

\item{roll_month}{See \code{?timechange::time_add}.}

\item{roll_dst}{See \code{?timechange::time_add}.}
}
\value{
A date, date-time, or other time-based vector.
}
\description{
A very fast method of adding time units to dates and date-times.
}
\details{
The methods are continuously being improved over time.
Date arithmetic should be very fast regardless of the timespan supplied.
Date-time arithmetic, specifically
when supplied days, weeks, months and years, is being improved.
}
