% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_count.R
\name{time_count}
\alias{time_count}
\title{\code{time_count} is deprecated}
\usage{
time_count(
  data,
  time = NULL,
  ...,
  time_by = NULL,
  from = NULL,
  to = NULL,
  .name = "{.col}",
  complete = FALSE,
  wt = NULL,
  name = NULL,
  sort = FALSE,
  .by = NULL,
  time_floor = FALSE,
  week_start = getOption("lubridate.week.start", 1),
  time_type = getOption("timeplyr.time_type", "auto"),
  roll_month = getOption("timeplyr.roll_month", "preday"),
  roll_dst = getOption("timeplyr.roll_dst", "NA"),
  as_interval = getOption("timeplyr.use_intervals", TRUE)
)
}
\arguments{
\item{data}{\bold{Deprecated}.}

\item{time}{\bold{Deprecated}.}

\item{...}{\bold{Deprecated}.}

\item{time_by}{\bold{Deprecated}.}

\item{from}{\bold{Deprecated}.}

\item{to}{\bold{Deprecated}.}

\item{.name}{\bold{Deprecated}.}

\item{complete}{\bold{Deprecated}.}

\item{wt}{\bold{Deprecated}.}

\item{name}{\bold{Deprecated}.}

\item{sort}{\bold{Deprecated}.}

\item{.by}{\bold{Deprecated}.}

\item{time_floor}{\bold{Deprecated}.}

\item{week_start}{\bold{Deprecated}.}

\item{time_type}{\bold{Deprecated}.}

\item{roll_month}{\bold{Deprecated}.}

\item{roll_dst}{\bold{Deprecated}.}

\item{as_interval}{\bold{Deprecated}.}
}
\description{
\code{time_count} is deprecated
}
