% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_ggplot.R
\name{time_ggplot}
\alias{time_ggplot}
\title{Quick time-series ggplot}
\usage{
time_ggplot(data, time, value, group = NULL, facet = FALSE, ...)
}
\arguments{
\item{data}{A data frame}

\item{time}{Time variable using \code{tidyselect}.}

\item{value}{Value variable using \code{tidyselect}.}

\item{group}{(Optional) Group variable using \code{tidyselect}.}

\item{facet}{When groups are supplied, should multi-series be
plotted separately or on the same plot?
Default is \code{FALSE}, or together.}

\item{...}{Further arguments passed to \code{geom_line()}.}
}
\description{
\code{time_ggplot()} is a neat way to quickly
plot aggregate time-series data.
}
\examples{
library(dplyr)
library(timeplyr)

# It's as easy as this
AirPassengers \%>\%
  ts_as_tibble() \%>\%
  time_ggplot(time, value)

# And this
EuStockMarkets \%>\%
  ts_as_tibble() \%>\%
  time_ggplot(time, value, group)

# An example using raw data

ebola <- outbreaks::ebola_sim$linelist
ebola \%>\%
  time_count(date_of_infection) \%>\%
  time_ggplot(date_of_infection, n)
}
