% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_seq_id.R
\name{time_seq_id}
\alias{time_seq_id}
\title{Generate a unique identifier for a regular time sequence with gaps}
\usage{
time_seq_id(
  x,
  time_by = NULL,
  g = NULL,
  time_type = c("auto", "duration", "period"),
  na_skip = TRUE
)
}
\arguments{
\item{x}{Date, datetime or numeric vector.}

\item{time_by}{Time unit. \cr
This signifies the granularity of the time data with which to measure gaps
in the sequence.
If your data is daily for example, supply \code{time_by = "days"}.
If weekly, supply \code{time_by = "week"}.
Must be one of the three:
\itemize{
\item string, specifying either the unit or the number and unit, e.g
\code{time_by = "days"} or \code{time_by = "2 weeks"}
\item named list of length one, the unit being the name, and
the number the value of the list, e.g. \code{list("days" = 7)}.
For the vectorized time functions, you can supply multiple values,
e.g. \code{list("days" = 1:10)}.
\item Numeric vector. If time_by is a numeric vector and x is not a date/datetime,
then arithmetic is used, e.g \code{time_by = 1}.
}}

\item{g}{Object used for grouping x.
This can for example be a vector or data frame.
\code{g} is passed directly to \code{collapse::GRP()}.}

\item{time_type}{If "auto", \code{periods} are used for
the time expansion when days, weeks, months or years are specified,
and \code{durations} are used otherwise.}

\item{na_skip}{Should \code{NA} values be skipped? Default is \code{TRUE}.}
}
\description{
A unique identifier is created every time a specied amount of
time has passed, or in the case of regular sequences, when there is a gap
in time.
}
\examples{
library(dplyr)
library(timeplyr)
library(lubridate)
# Weekly sequence, with 2 gaps in between
x <- time_seq(today(), length.out = 10, time_by = "week")
x <- x[-c(3, 7)]
time_seq_id(x, time_by = "week")
time_gaps(x, time_by = "week") # Time gaps
}
