% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeplyr-package.R
\docType{package}
\name{timeplyr-package}
\alias{timeplyr}
\alias{timeplyr-package}
\title{timeplyr: Fast Tidy Tools for Date and Date-Time Manipulation}
\description{
A framework for handling raw date & datetime data \cr
using tidy best-practices from the tidyverse, the efficiency of data.table,
and the speed of collapse.

You can learn more about the tidyverse,
data.table and collapse using the links below

\href{https://www.tidyverse.org/learn/}{tidyverse}

\href{https://CRAN.R-project.org/package=data.table}{data.table}

\href{https://sebkrantz.github.io/collapse/articles/collapse_intro.html}{collapse}
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/NicChr/timeplyr/issues}
}

}
\author{
\strong{Maintainer}: Nick Christofides \email{nick.christofides.r@gmail.com} (\href{https://orcid.org/0000-0002-9743-7342}{ORCID})

}
\keyword{internal}
