% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeplyr.R
\docType{package}
\name{timeplyr-package}
\alias{timeplyr}
\alias{timeplyr-package}
\title{timeplyr: Date and datetime extension to dplyr}
\description{
A framework for handling raw date & datetime data \cr
using tidy best-practices from the tidyverse, the efficiency of data.table,
and the speed of collapse.

You can learn more about the tidyverse,
data.table and collapse using the links below

\href{https://www.tidyverse.org/learn/}{tidyverse}

\href{https://cran.r-project.org/web/packages/data.table/vignettes/datatable-intro.html}{data.table}

\href{https://sebkrantz.github.io/collapse/articles/collapse_intro.html}{collapse}
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/NicChr/timeplyr/issues}
}

}
\author{
\strong{Maintainer}: Nick Christofides \email{nick.christofides.r@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
