% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timespan.R
\name{timespan}
\alias{timespan}
\title{Timespans}
\usage{
timespan(units, num = 1L, ...)
}
\value{
A \link{timespan} object.
}
\description{
Timespans
}
\examples{
library(timeplyr)

timespan("week")
timespan("day")
timespan("decade")

# Multiple units of time

timespan("10 weeks")
timespan("1.5 hours")
}
